requirements.txt
cgitb
numpy
pandas
Pillow
torch
scikit-learn
transformers
torchvision